fix(client): extract client-safe redirect URL sanitizer to resolve next/server bundler defect
PR: #<PR_NUMBER>
Architectural Precedent: ADR-002 (Client Layer Boundaries), ADR-004 (Environment Boundaries)

### Summary
Extracted `getSafeRedirectUrl` from `app/lib/security/middleware/redirect-policy.ts` into a standalone, client-compatible module at `app/lib/security/redirect-url.ts`.

Previously, `ClerkSignInWidget.tsx` ("use client") and `AuthCallbackPage` ("use client") imported `getSafeRedirectUrl` directly from `redirect-policy.ts`. Because `redirect-policy.ts` imports `next/server` (`NextRequest`, `NextResponse`), Next.js turbopack/webpack client bundlers threw a critical build failure (`"next/server" is server-only`).

### Key Changes
- **Client-Safe Sanitizer Module (`app/lib/security/redirect-url.ts`)**: Created a dedicated module for `getSafeRedirectUrl(target)` with zero `next/server` imports. Retains strict open-redirect protection (relative paths, `*.buildmarket.app` origins, configured app/admin/verification origins, loopback hosts) using `env` configuration from `@/app/lib/infrastructure/env`.
- **Middleware Policy Facade (`app/lib/security/middleware/redirect-policy.ts`)**: Re-exports `getSafeRedirectUrl` from `redirect-url.ts` to preserve backwards compatibility for server/middleware callers while eliminating unused `env` imports.
- **Client & Page Imports (`ClerkSignInWidget.tsx`, `app/auth-callback/page.tsx`, `app/sign-in/[[...sign-in]]/page.tsx`)**: Updated `getSafeRedirectUrl` imports to consume `@/app/lib/security/redirect-url`.
- **Vitest Unit Test Suites**: Created `__tests__/lib/redirect-url.test.ts` (8 isolated unit tests) and updated `__tests__/lib/redirect-policy.test.ts` (17 unit tests).
- **Changelog Governance**: Documented fix under `apps/client/docs/CHANGELOG.md` § `[Unreleased]`.

### Verification & Artifact Evidence
- `pnpm -C apps/client exec vitest run __tests__/lib/redirect-policy.test.ts __tests__/lib/redirect-url.test.ts` -> 25 tests passed across 2 test files (0 failures).
- `pnpm -C apps/client run check-env-contract` -> Exit code 0 (`[env-contract] OK: .env.example covers all process.env keys`).

Affected Files:
- `apps/client/app/lib/security/redirect-url.ts` [NEW]
- `apps/client/__tests__/lib/redirect-url.test.ts` [NEW]
- `apps/client/app/lib/security/middleware/redirect-policy.ts`
- `apps/client/components/auth/ClerkSignInWidget.tsx`
- `apps/client/app/auth-callback/page.tsx`
- `apps/client/app/sign-in/[[...sign-in]]/page.tsx`
- `apps/client/__tests__/lib/redirect-policy.test.ts`
- `apps/client/docs/CHANGELOG.md`
