+++ CHANGELOG.md
@@ -792,7 +792,7 @@ See `docs/operations/professional-onboarding-observability-runbook.md` §8.
 - **TypeScript/Build Pipeline**: Resolved CI and Vercel build failures caused by the native compiler upgrade to TypeScript `7.0.2`. Next.js 16.2.6 expects a programmatic `typescript/lib/typescript.js` file, which is missing from TS7's native binary bundle. Added a workspace `postinstall` script ([patch-typescript.mjs](file:///c:/Users/User/build-market/scripts/patch-typescript.mjs)) that patches the `typescript` package in `node_modules` by bridging it to standard `@typescript/typescript6` compiler APIs. This resolves programmatic compilation checks by Next.js and `tsx` in CI while retaining the fast native `tsc` binary for CLI compilation.
 - **Dependencies**: Added `tsx` to `apps/client` devDependencies and cataloged it monorepo-wide to ensure clean execution under CI.
 
-### Changed()
+### Changed (Typescript)
 
 - **TypeScript**: Configured `@typescript-eslint` compiler compatibility routing via `pnpm` workspace overrides ([pnpm-workspace.yaml](file:///c:/Users/User/build-market/pnpm-workspace.yaml)), mapping `typescript` dependency inside `@typescript-eslint/*`, `@build/eslint-config`, and `eslint-config-next` packages to resolve to `@typescript/typescript6`. This enables using TypeScript 7's Go-powered native compiler for building while keeping `typescript-eslint` routed through the backward compatible TypeScript 6 wrapper for static analysis.
 - **TypeScript**: Removed deprecated/removed `"baseUrl": "."` option from `apps/client/tsconfig.json` ([tsconfig.json](file:///c:/Users/User/build-market/apps/client/tsconfig.json)) to satisfy TypeScript 7 compiler rules.
