+++ apps/client/.env.development
@@ -15,7 +15,8 @@ NEXT_PUBLIC_APP_URL=http://localhost:3500
 NEXT_PUBLIC_API_URL=http://localhost:3500/api
 NEXT_PUBLIC_SEARCH_SERVICE_URL=http://localhost:3005
 NEXT_PUBLIC_ADMIN_APP_URL=http://localhost:3005
-
+NEXT_PUBLIC_VERIFICATION_APP_URL=http://localhost:3501
+NEXT_PUBLIC_VERIFICATION_OPS_URL=http://localhost:3501
 # Clerk Routes
 # FIX: removed the leading space and incorrect localhost URL that was here.
 # NEXT_PUBLIC_CLERK_FRONTEND_API must match your Clerk dashboard instance URL
