+++ apps/client/.env.example
@@ -102,6 +102,8 @@ NEXT_PUBLIC_APP_URL=http://localhost:3500
 NEXT_PUBLIC_API_URL=http://localhost:3500/api
 NEXT_PUBLIC_SEARCH_SERVICE_URL=http://localhost:3005
 NEXT_PUBLIC_ADMIN_APP_URL=http://localhost:3005
+NEXT_PUBLIC_VERIFICATION_APP_URL=http://localhost:3501
+NEXT_PUBLIC_VERIFICATION_OPS_URL=http://localhost:3501
 
 # Server-side canonical URL (used by worker notifications)
 APP_URL=http://localhost:3500
