+++ apps/client/.env.vercel.example
@@ -26,7 +26,8 @@ NEXT_PUBLIC_CLERK_SIGN_UP_FALLBACK_REDIRECT_URL=/onboarding
 NEXT_PUBLIC_CLERK_AFTER_SIGN_IN_URL=/auth-callback
 NEXT_PUBLIC_CLERK_AFTER_SIGN_UP_URL=/onboarding
 NEXT_PUBLIC_ADMIN_APP_URL=https://build-market-ebon.vercel.app/admin
-
+NEXT_PUBLIC_VERIFICATION_APP_URL=https://build-market-ebon.vercel.app/verification
+NEXT_PUBLIC_VERIFICATION_OPS_URL=https://build-market-ebon.vercel.app/verification-ops
 # Auth [REQUIRED]
 # AUTH_SECRET=your_production_auth_secret_base64
