+++ apps/client/docs/CHANGELOG.md
@@ -32,7 +32,9 @@ This format is based on Keep a Changelog and uses semantic categories:
 
 - **Satellite Cross-Domain Auth Redirects (`app/sign-in/`, `app/auth-callback/`, `app/lib/security/middleware/redirect-policy.ts`)**: fixed stuck "Welcome Back" sign-in card loop when navigating from satellite domains (`verification.buildmarket.app`, `admin.buildmarket.app`) to `https://buildmarket.app/sign-in?redirect_url=...`:
   - **Environment Configuration (`env.ts`)**: declared `NEXT_PUBLIC_VERIFICATION_APP_URL` / `NEXT_PUBLIC_VERIFICATION_OPS_URL` in `urls` group and wired `verificationAppUrl` property into `buildEnvConfig()` with `http://localhost:3501` dev fallback.
-  - **Redirect Target Sanitizer (`redirect-policy.ts`)**: implemented `getSafeRedirectUrl(target)` validating relative paths, `*.buildmarket.app` subdomains, configured app/admin/verification origins, and local loopback hosts (`localhost`, `127.0.0.1`) while blocking open-redirect attack vectors (`//evil.com`, `javascript:`, untrusted domains).
+  - **Client-Safe Redirect Sanitizer (`app/lib/security/redirect-url.ts`)**: extracted `getSafeRedirectUrl` into a standalone, client-compatible module with zero `next/server` dependencies to prevent Next.js build errors when imported by client components (`ClerkSignInWidget.tsx`, `AuthCallbackPage`). `redirect-policy.ts` re-exports `getSafeRedirectUrl` for server/middleware compatibility.
+  - **Server Page & Client Component Imports (`app/sign-in/[[...sign-in]]/page.tsx`, `components/auth/ClerkSignInWidget.tsx`, `app/auth-callback/page.tsx`)**: updated imports to read directly from `app/lib/security/redirect-url.ts`.
+  - **Redirect URL Test Suite (`__tests__/lib/redirect-url.test.ts`)**: added dedicated Vitest unit test suite validating redirect URL sanitization in isolation without server-only dependencies.
 
   - **Server Page Pre-Redirection (`app/sign-in/[[...sign-in]]/page.tsx`)**: checked server-side `auth()` in `SignInPage` to immediately redirect signed-in users to `safeRedirectUrl` or `/auth-callback`.
   - **Client Widget & Auth Callback (`ClerkSignInWidget.tsx`, `app/auth-callback/page.tsx`)**: added client-side `useUser` hydration redirect and updated `AuthCallbackPage` to honor `redirect_url` search parameter post-sign-in.
